+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for `snake_case` (underscore style) arguments. (#6)
+- A `DeprecationWarning` is now shown when using `Choices` instead of `Literal`.
 
 ## [0.3.0]
 ### Added